+++ framework_src/native_library/FreSwiftExampleANE/FreSwift/Geom/FrePointSwift.swift
@@ -13,9 +13,6 @@
  limitations under the License.*/
 
 import Foundation
-#if canImport(Cocoa)
-    import Cocoa
-#endif
 /// :nodoc:
 public class FrePointSwift: FreObjectSwift {
     override public init(freObject: FREObject?) {