+++ framework_src/native_library/FreSwiftExampleANE/FreSwiftExampleANE/SwiftController.swift
@@ -79,22 +79,27 @@ public class SwiftController: NSObject {
     }
     
     func runArrayTests(ctx: FREContext, argc: FREArgc, argv: FREArgv) -> FREObject? {
-        trace("***********Start Array test NEW ***********")
+        trace("***********Start Array test ***********")
         
         guard argc > 0, let inFRE0 = argv[0] else {
             return nil
         }
         
         let airArray: FREArray = FREArray(inFRE0)
+        
         do {
             let airArrayLen = airArray.length
             
             trace("Array passed from AIR:", airArray.value)
             trace("AIR Array length:", airArrayLen)
+            for fre in airArray {
+                trace("iterate over FREArray", Int(fre) ?? "unknown")
+            }
             
-            if let itemZero = try Int(airArray.at(index: 0)) {
+            if let itemZero = Int(airArray[0]) { //get using brackets with FREObject
                 trace("AIR Array elem at 0 type:", "value:", itemZero)
                 try airArray.set(index: 0, value: 56)
+                airArray[1] = 123.toFREObject() //set using brackets with FREObject
                 return airArray.rawValue
             }
             
@@ -102,7 +107,7 @@ public class SwiftController: NSObject {
             _ = e.getError(#file, #line, #column)
         } catch {
         }
-        
+
         return nil
         
     }